requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-google-genai
google-generativeai
python-dotenv
requests
pydantic
fastmcp
newsapi-python
anyio